ABC→MusicXML の互換パーサー導入と実運用記法対応（M:C/C|・broken rhythm・tie・x休符・和音・連符）
### 背景
既存の `ABC -> MusicXML` 変換は、実運用で多く見られるABC譜（`abcjs`/`abcm2ps` 系）に対して、以下のような記法で変換エラーになりやすい状態でした。

- `M:C` / `M:C|`
- `A > B` のような空白入り broken rhythm
- `x` 休符
- `"(3..."` 連符
- `"[...]"` 和音
- インラインテキスト（`"..."`）
- 一部の装飾・補助記号の混在

### 変更概要
#### 1. 互換パーサーを新規追加
- `docs/music/src/common/ts/abc-compat-parser.ts`（+ JS生成物）
- `AbcCommon` を利用し、`ABC 2.1` をベースに「実データ互換寄り」の解析を実装
- `docs/music/src/abc-to-musicxml/ts/main.ts` から旧インライン実装を分離し、`AbcCompatParser` 呼び出しに置換

#### 2. 対応記法の拡張
- ヘッダ: `X/T/C/M/L/K/V`, `%%score`
- メーター: `M:C`, `M:C|` を解釈
- 音価: broken rhythm `>`/`<`（空白入り含む）
- tie: `-`
- 休符: `z/Z` に加え `x/X`
- インライン文字列 `"..."` は警告付きスキップ
- 装飾 `!...!`, `+...+` は警告付きスキップ
- 連符: `(3` を含む tuplet 記法 `n[:q[:r]]` の音価スケーリング
- 和音: `[ ... ]` を同時発音として展開

#### 3. MusicXML出力の拡張
- `docs/music/src/common/ts/musicxml-writer-common.ts`
- 和音ノートに `<chord/>` を出力

#### 4. ビルド導線の更新
- `scripts/build-music.mjs` に `abc-compat-parser` を追加
- `docs/music/abc-to-musicxml-src.html` / `docs/music/abc-to-musicxml.html` へスクリプト読み込みを追加

#### 5. 仕様ドキュメント追加
- `docs/music/abc-compat-parser-ebnf.md`
- 互換実装の対象範囲を EBNF（draft）として明文化

### テスト
- 追加: `docs/music/tests/abc-compat-parser.test.js`
  - spaced broken rhythm
  - `M:C` + mixed token
  - transpose inference (`in A`)
  - tuplet/chord
  - standalone octave marker互換ケース
- 既存更新: `docs/music/tests/abc-to-musicxml-main.test.js`
  - 新パーサー読み込み
  - tuplet/chord統合テスト追加

### 期待効果
- 実運用ABC譜（特にダンス曲系・伴奏譜）での変換失敗を大幅に低減
- `ABC -> MusicXML` 変換の保守性を向上（パーサー責務を分離）
- 今後の互換拡張を、仕様（EBNF）とテストで継続的に育てられる状態に改善
