+++ docs/music/abc-to-musicxml-src.html
@@ -213,6 +213,7 @@ C D E F | G A B c |</textarea>
     <script src="src/common/js/musicxml-synth-schedule-common.js"></script>
     <script src="src/common/js/music-synth-common.js"></script>
     <script src="src/common/js/musicxml-writer-common.js"></script>
+    <script src="src/common/js/abc-compat-parser.js"></script>
     <script src="src/abc-to-musicxml/js/main.js"></script>
 </body>
 </html>
